requirements.txt
tensorflow>=2.0.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
sympy>=1.8.0
pillow>=8.0.0
flask>=2.0.0
